feat: Add env of comment entity
Comment and its environment:
DB contains orders which is received from API. Every user should have an
opportunity to mark them by comment. In order to protect the last of
from removing I decided to separate it into an own entity. So, it will
give to users with different rights a view with or without comments
since this is one of the important requirements for the program.