fix(entity): fix LazyInitializationException
All entities were reformatted at their annotations. The working way is
to use @EntityGraph as it is used in Company and CompanyRepository.
Program contains two equal methods without little differences in
OrderJob and OrderServiceImpl. The first works with @Scheduled
annotation and loads for itself all service classes. The second
mystically loses the connection between entities with or without
@Transactional anno. Methods which work only with local DB without
sending request to API do it in right way.