feat: Add index fields, edit request timeout
Order, Document
- In order to get better and faster database search decided to make uid
fields in both entities indexed. It becomes possible also because of
editing request timeout. As we know, indexing increases the time it
takes to write, update, and delete data in a table, because the database
must update the index on every change. But we should have a recommended
delay between request about 10 sec. So, I rewrote the delay time for the
remainder between 10 seconds and the time to complete the request

OrderServiceImpl
- Change request timeout

OrderJob
- Edit logging