fix: Fix update method test
The reason of unreachable body of the response was in service and in
process of mocking it. First of all, it used in two ways in one method:
as mocked and not. Secondly, the service methods that use Retrofit
should have a baseUrl, which is a var in application.yml. So, the best
way was to split every service class in order to separate methods,
interface and, of course, the logic of work with Retrofit.